requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-multipart==0.0.6
pillow==10.0.1
pydantic==2.4.2
python-dotenv==1.0.0
libphonenumber==8.13.25
python-jose[cryptography]==3.3.0
email-validator==2.1.0 